extract AccessMode and StorageType to their own file
