+++ backend/de.metas.adempiere.adempiere/base/src/main/java/org/adempiere/archive/api/AccessMode.java
@@ -1,0 +1,11 @@
+package org.adempiere.archive.api;
+
+/**
+ * AccessMode - from where the archive is accessed (client side, server side)
+ */
+public enum AccessMode
+{
+	CLIENT,
+	SERVER,
+	ALL,
+}
