+++ backend/de.metas.adempiere.adempiere/base/src/main/java/org/adempiere/archive/api/impl/ArchiveStorageFactory.java
@@ -27,7 +27,9 @@ import de.metas.logging.LogManager;
 import de.metas.util.Services;
 import lombok.NonNull;
 import lombok.Value;
+import org.adempiere.archive.api.AccessMode;
 import org.adempiere.archive.api.IArchiveStorageFactory;
+import org.adempiere.archive.api.StorageType;
 import org.adempiere.archive.spi.IArchiveStorage;
 import org.adempiere.archive.spi.impl.DBArchiveStorage;
 import org.adempiere.archive.spi.impl.FilesystemArchiveStorage;
