+++ backend/de.metas.document.archive/de.metas.document.archive.base/src/main/java/de/metas/document/archive/rpl/requesthandler/ArchiveSetDataHandlerConverter.java
@@ -26,6 +26,7 @@ package de.metas.document.archive.rpl.requesthandler;
 import java.util.Properties;
 
 import org.adempiere.archive.api.IArchiveStorageFactory;
+import org.adempiere.archive.api.StorageType;
 import org.adempiere.archive.spi.IArchiveStorage;
 import org.adempiere.model.InterfaceWrapperHelper;
 import org.compiere.model.MTable;
@@ -48,7 +49,7 @@ public class ArchiveSetDataHandlerConverter implements Converter<I_AD_Archive, I
 		final byte[] data = tempArchive.getBinaryData();
 
 		final Properties ctx = InterfaceWrapperHelper.getCtx(tempArchive);
-		final IArchiveStorage fsStorage = Services.get(IArchiveStorageFactory.class).getArchiveStorage(ctx, IArchiveStorageFactory.StorageType.Filesystem);
+		final IArchiveStorage fsStorage = Services.get(IArchiveStorageFactory.class).getArchiveStorage(ctx, StorageType.Filesystem);
 		fsStorage.setBinaryData(tempArchive, data);
 		InterfaceWrapperHelper.save(tempArchive);
